Direct upload of shooter game try
This game is written with python's pygame library and made use of modules and etc. to be closer to real life applications and decency of professionality. Another important is that, this game is an appearinf-g of curiosity and enthusiasm of self improvement. It lacks of concepts, codes and etc. It has need of getting bug free.